Id-Handling for VideoConversion  updated
The autoincremented id is changed to the sourceId because this id is
known to the source system (Lecture2Go) and calls can easily be made.

For every sourceid there can only be one conversion, old or current
conversion processes for an id will be cleaned and the database entry
reseted.